requirements.txt
loguru~=0.7.0
accelerate~=1.5.2
diffusers~=0.32.2
gguf~=0.14.0
huggingface-hub~=0.29.3
protobuf~=6.30.1
python-dotenv~=1.0.1
safetensors~=0.5.3
sentencepiece~=0.2.0
tokenizers~=0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers~=4.50.0
flask==3.0.2
flask-wtf==1.2.1
flask-cors==4.0.0
Werkzeug==3.0.1
requests==2.31.0
